Move fields combination to API
