+++ main/src/main/java/com/kenshoo/pl/entity/internal/validators/FieldsCombinationValidationAdapter.java
@@ -1,7 +1,6 @@
 package com.kenshoo.pl.entity.internal.validators;
 
 import com.kenshoo.pl.entity.*;
-import com.kenshoo.pl.entity.internal.FieldsCombination;
 import com.kenshoo.pl.entity.spi.FieldsCombinationValidator;
 import java.util.stream.Stream;
